Moved helper functions to testutils, so that they can be used by both c2s and s2c. Used those new helper functions to fix a freezing error in s2c, where the s2c test pauses until it hears back from the packet trace engine, and so if the packet trace engine fails to start, then the s2c test just freezes forever.
